% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_otrh.R
\name{compute_orth}
\alias{compute_orth}
\title{Data Frame içindeki Kelimeler için Dilsel Metrikleri Hesaplama}
\usage{
compute_orth(df_input, word_col, lang = "turkish", bigram.type = "mean")
}
\arguments{
\item{df_input}{The main data frame containing the words..}

\item{word_col}{The name of the column containing the words (as a string, e.g., "word").}

\item{lang}{The language to be used for the dictionary (e.g., 'turkish').}

\item{bigram.type}{Calculation type ('mean' or 'summed').}
}
\description{
Data Frame içindeki Kelimeler için Dilsel Metrikleri Hesaplama
}
