% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_paragraphs.R
\name{count_paragraphs}
\alias{count_paragraphs}
\title{Number of paragraphs in the text}
\usage{
count_paragraphs(text.path, title = TRUE)
}
\arguments{
\item{text.path}{The path of the .txt file.}

\item{title}{logical value indicating whether it is a header (TRUE,FALSE)}
}
\value{
Returns the input text as number.
}
\description{
This script finds the number of paragraphs in the text file 
located at the given path..
}
\examples{
\dontrun{
  count_paragraphs(text.path = "texts/text1.txt", title = TRUE)
}
}
