% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_syl.R
\name{count_syl}
\alias{count_syl}
\title{Number of syllables in the word}
\usage{
count_syl(words)
}
\arguments{
\item{word}{A vector includes words.}
}
\value{
Returns the input word as number.
}
\description{
This script finds the number of syllable in any word
}
\examples{
\dontrun{
  count_syl(c('masa','arkadaş','ağaç'))
}
}
