% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_syl.R
\name{syllabify_naive}
\alias{syllabify_naive}
\title{Find syllables in any word}
\usage{
syllabify_naive(word)
}
\arguments{
\item{word}{A vector includes words.}
}
\value{
Returns the input word as syllabified word.
}
\description{
This script finds the syllables in any word
}
\examples{
\dontrun{ 
syllabify_naive(c('masa','arkadaş','ağaç'))
}
}
