% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtSimilarity.R
\name{txtSimilarity}
\alias{txtSimilarity}
\title{Text similarity tests}
\usage{
txtSimilarity(text.paths, pgraphs, titles = c(FALSE, FALSE))
}
\arguments{
\item{text.paths}{The paths to the .txt files.You should combine the variables one by one using c() to form vectors. eg., c(tex1.path,tex2.path)}

\item{pgraphs}{Number of paragraphs in the texts (excluding the title).You should combine the variables one by one using c() to form vectors. eg., c(3,2)}

\item{titles}{logical value indicating whether they are headers. You should combine the variables one by one using c() to form vectors. eg., c(TRUE,FALSE)}
}
\value{
Returns the input texts as a data frame.
}
\description{
Identifies similarities between two texts written in txt format.Provides Jaccard Similarity and Cousine Similarity values.
}
\examples{
\dontrun{
  txt_similarity(text.paths=c(text.path='texts/text1.txt', text.path='texts/text1.txt'), pgraph=c(2,3),titles=c(TRUE,FALSE))
}
}
