% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wordlist.R
\name{make_wordlist}
\alias{make_wordlist}
\title{Converting text to a word list}
\usage{
make_wordlist(
  text.path,
  pgraph,
  title = FALSE,
  bigram.type = "mean",
  lang = "turkish"
)
}
\arguments{
\item{text.path}{The path to the .txt file.}

\item{pgraph}{Number of paragraphs in the text (excluding the title)}

\item{title}{logical value indicating whether it is a header (TRUE, FALSE)}

\item{bigram.type}{Whether the bigram value of the word should be the mean 
or the sum value (mean, summed)}

\item{lang}{The language in which the text is written.Currently only 
available for some language's texts. 
Parameters: basque,dutch,english,french,german,italian,polish,serbian_cyrillic, 
serbian_latin, spanish,turkish,vietnamese}
}
\value{
Returns the input text as a data frame.
}
\description{
It provides the number of words and sentences in a text and 
the number of letters and syllables for each word, the OLD20 value 
according to the selected language, and the bigram value in a table..
}
\examples{
make_wordlist(text.path='texts/text1.txt',pgraph=2,title=TRUE, 
                        bigram.type='mean',lang='turkish')
}
