% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable.R
\name{readable}
\alias{readable}
\title{Readability tests}
\usage{
readable(df, returnType = "long")
}
\arguments{
\item{df}{the data frame returned by the make_wordlist() function.}

\item{returnType}{Determines whether the data frame to be returned will 
be wide or long}
}
\value{
Returns data frame .
}
\description{
It processes the data frame returned by the make_wordlist() 
function and provides readability measurements for the text. The syllable 
column in the data frame returned by the make_wordlist() function must 
contain values. It also includes readability criteria proposed for Turkish.
}
\examples{
readable(df=df,returnType='long')
}
