% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_wordlist.R
\name{sum_wordlist}
\alias{sum_wordlist}
\title{Removing duplicates from the word list}
\usage{
sum_wordlist(df)
}
\arguments{
\item{df}{the data frame returned by the make_wordlist() function.}
}
\value{
Returns data frame .
}
\description{
It processes the data frame returned by the make_wordlist() 
function and provides readability measurements for the text. The syllable 
column in the data frame returned by the make_wordlist() function must 
contain values. It also includes readability criteria proposed for Turkish.
}
\examples{
sum_wordlist(df=df)
}
